+++ CHANGELOG.md
@@ -28,7 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
-- ***
+- Updates encodeUrl to take in the queryString state and defines its output as part of the return statement
 
 ### Fixed